socketndpi/requirements.txt
dpkt>=1.9.7.2
prometheus-client
python-dotenv
SQLAlchemy